Adds quick and dirty accessControl -> - /post can only be accessed if a postUri is set as a param in the url - /settings can only be accessed when a user is logged in - /signup is only acccessible if there is no user logged in or the logged in user is anonymous
If the route is currently not allowed a redirect to the start page will be initiated
